# chore: optimize sealed block read