# Benchmarks
## DB reads
```
❯ SKIP_GUEST_BUILD=1 cargo bench --bench db_operations_bench -p sov-db
   Compiling sov-db v0.7.4 (/home/rakan/Work/Chainway/citrea/crates/sovereign-sdk/full-node/db/sov-db)
    Finished `bench` profile [optimized] target(s) in 16.29s
     Running benches/db_operations_bench.rs (target/release/deps/db_operations_bench-86cc56588f63987b)
Gnuplot not found, using plotters backend
single_read/100         time:   [414.72 ns 418.32 ns 422.81 ns]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
single_read/1000        time:   [434.84 ns 436.45 ns 437.99 ns]
single_read/10000       time:   [428.23 ns 430.12 ns 432.02 ns]

sequential_reads/10     time:   [4.1761 µs 4.1948 µs 4.2144 µs]
sequential_reads/100    time:   [45.910 µs 45.965 µs 46.046 µs]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe

random_reads/100        time:   [47.263 µs 47.443 µs 47.654 µs]
Found 18 outliers among 100 measurements (18.00%)
  5 (5.00%) low severe
  4 (4.00%) low mild
  3 (3.00%) high mild
  6 (6.00%) high severe
random_reads/1000       time:   [47.982 µs 48.075 µs 48.149 µs]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild

miss_reads/100          time:   [445.66 ns 447.51 ns 449.34 ns]
miss_reads/1000         time:   [462.58 ns 464.81 ns 467.00 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) low mild
miss_reads/10000        time:   [471.90 ns 473.85 ns 475.81 ns]
```

## seal_slow calls
```
❯ SKIP_GUEST_BUILD=1 cargo bench --bench seal_slow_bench -p citrea-evm
   Compiling citrea-evm v0.7.4 (/home/rakan/Work/Chainway/citrea/crates/evm)
    Finished `bench` profile [optimized] target(s) in 15.63s
     Running benches/seal_slow_bench.rs (target/release/deps/seal_slow_bench-b2a38c795d7fce6a)
Gnuplot not found, using plotters backend
seal_slow_empty_block   time:   [1.9332 µs 1.9356 µs 1.9380 µs]
                        change: [-1.5617% -1.4579% -1.3647%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) low mild

seal_slow_tx_count/0    time:   [1.9669 µs 1.9709 µs 1.9742 µs]
                        change: [-1.4410% -1.0831% -0.7103%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  9 (9.00%) low severe
  6 (6.00%) low mild
seal_slow_tx_count/10   time:   [1.9659 µs 1.9724 µs 1.9786 µs]
                        change: [-4.6087% -4.1943% -3.7801%] (p = 0.00 < 0.05)
                        Performance has improved.
seal_slow_tx_count/100  time:   [1.9706 µs 1.9779 µs 1.9852 µs]
                        change: [-5.0616% -4.7239% -4.4268%] (p = 0.00 < 0.05)
                        Performance has improved.
seal_slow_tx_count/500  time:   [1.9589 µs 1.9674 µs 1.9757 µs]
                        change: [-3.9383% -3.5125% -3.0862%] (p = 0.00 < 0.05)
                        Performance has improved.
seal_slow_tx_count/1000 time:   [1.9701 µs 1.9788 µs 1.9870 µs]
                        change: [-4.0346% -3.6302% -3.2452%] (p = 0.00 < 0.05)
                        Performance has improved.

seal_slow_extra_data/0  time:   [1.9637 µs 1.9713 µs 1.9787 µs]
                        change: [-1.2122% -0.8563% -0.4834%] (p = 0.00 < 0.05)
                        Change within noise threshold.
seal_slow_extra_data/32 time:   [1.9887 µs 1.9978 µs 2.0066 µs]
                        change: [-1.9671% -1.5923% -1.2045%] (p = 0.00 < 0.05)
                        Performance has improved.
seal_slow_extra_data/64 time:   [1.9777 µs 1.9900 µs 2.0028 µs]
                        change: [-3.3267% -2.8028% -2.2540%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
seal_slow_extra_data/128
                        time:   [2.2201 µs 2.2299 µs 2.2402 µs]
                        change: [-2.7223% -2.3110% -1.9102%] (p = 0.00 < 0.05)
                        Performance has improved.
seal_slow_extra_data/256
                        time:   [2.4115 µs 2.4223 µs 2.4337 µs]
                        change: [-5.2321% -4.8444% -4.4402%] (p = 0.00 < 0.05)
                        Performance has improved.
```

# Conclusion

Based on this, this PR is a demonstration that DB reads are around 4 times faster so we don't need this optimization of DB reads.

The reason why we thought this before was to make sure we don't need DB reads to collect mempool maintenance information. However, trying to remove the one DB read means that we have to read data PRIOR to the cache / working_set being committed. In this case, we won't have access to `SealedBlock` since it hasn't been sealed before committing it to DB.

## Linked Issues
- Fixes #2907