@rakanalh 2025-10-01
> Can you try with `seal_unchecked` instead since we already have the block hash ?

block hash != evm hash

`seal_unchecked` requires the evm hash which we don't have.