@rakanalh 2025-10-01
> since we call finalize_hook before we call the build reth block functions, can't we just read the sealed block from the EVM rpc functions?

We call finalize AFTER... The whole idea is that we want to prevent DB read and read it from the working set... if we call finalize, then we commit and reset the working_set which means that we lose the cache resulting in a DB read.
Calling the RPC functions would do the same.
[3:05](https://chainway.slack.com/archives/D06RZHR5SBY/p1759320322512789)
Basically means that we keep the current approach of having a single DB read instead of using seal_slow